requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
blinker==1.9.0
Bottleneck==1.4.2
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
db-dtypes==1.4.2
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
distro==1.9.0
easyocr==1.7.2
exceptiongroup==1.2.2
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
frozenlist==1.6.0
fsspec==2025.3.2
gcsfs==2025.3.2
google-api-core==2.24.2
google-auth==2.39.0
google-auth-oauthlib==1.2.1
google-cloud-bigquery==3.31.0
google-cloud-core==2.4.3
google-cloud-documentai==3.3.0
google-cloud-documentai-toolbox==0.14.2a0
google-cloud-storage==3.1.0
google-cloud-vision==3.10.1
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
h5py==3.13.0
httpcore==1.0.8
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
imageio==2.37.0
immutabledict==4.2.1
importlib_metadata==8.6.1
intervaltree==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
langdetect==1.0.9
lazy_loader==0.4
llvmlite==0.43.0
lxml==5.3.2
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.4.3
networkx==3.2.1
ninja==1.11.1.4
numba==0.60.0
numexpr==2.10.2
numpy==1.26.4
oauthlib==3.2.2
onnx==1.17.0
openai==1.74.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==24.2
pandas==2.2.3
pandas-gbq==0.28.0
pikepdf==8.15.1
pillow==10.4.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
pyarrow==15.0.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pydantic==2.11.3
pydantic_core==2.33.1
pydata-google-auth==1.9.1
pypdfium2==4.30.1
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-doctr==0.10.0
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
scikit-image==0.24.0
scipy==1.13.1
shapely==2.0.7
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sympy==1.13.3
tabulate==0.9.0
tifffile==2024.8.30
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.20.0
zipp==3.21.0
